Changed something

+++ Assets/Scripts/Vegetables/CabbageClass.cs
@@ -9,7 +9,7 @@ public class CabbageClass : VegetableClass
         
     }
 
-    public CabbageClass(int health, float damage, float speed, string prefab, GameObject gameObject)
+    public CabbageClass(float health, float damage, float speed, string prefab, GameObject gameObject)
         : base(health, damage, speed, prefab, gameObject)
     {
